+++ PyParagraph/main.py
@@ -13,9 +13,10 @@ while keep_going == 'y':
     print("|  Analysis are saved as <name of file>.txt in \Analysis folder  |")
     print("|                 Make sure file names are unique                |")
     print("==================================================================")
-    #file = input("What's the name of the file you want to analyze? (in Resources folder):")
-    #for testing purposes, switch comment above and below when testing is done
-    file = 'testing.txt'
+    
+    #for testing purposes, switch comment below when testing is done
+    file = input("What's the name of the file you want to analyze? (in Resources folder):")
+    #file = 'testing.txt'
 
     #checking if the file exists
     txtpath = os.path.join('Resources', file)
@@ -90,7 +91,6 @@ while keep_going == 'y':
         for line in analysis:
             print (line, end = " ")
 
-    keep_going = 'n'
-    #keep_going = input ("Do you want to analyze another file? (y/n): ")
-
-print (sentences)
+    #for testing purposes, switch comment below when testing is done
+    #keep_going = 'n'
+    keep_going = input ("Do you want to analyze another file? (y/n): ")