+++ PyParagraph/changelog.md
@@ -27,4 +27,4 @@
 - tested script with testing paragraph
     - discrepancies between test results provided and script
     - Script, MS-word and [WordCounter](https://wordcounter.net/) counts 120 words in the test paragraph, script is correct
-    
+- cleaned up the testing code